% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molpathR-package.R
\docType{package}
\name{molpathR-package}
\alias{molpathR}
\alias{molpathR-package}
\title{molpathR: Molecular Pathology Data Platform for Clinical NGS Integration}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A unified platform for ingesting, linking, querying, and visualizing molecular pathology data from heterogeneous sources including NGS files (FASTQ, BAM, VCF), XML variant interpretation reports, structured pathology PDF reports, pathology and clinical information systems, and survival outcome data. Provides a searchable in-memory database and an interactive Shiny application for clinical exploration and cohort analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cttir/molpathR}
  \item \url{https://cttir.github.io/molpathR/}
  \item Report bugs at \url{https://github.com/cttir/molpathR/issues}
}

}
\author{
\strong{Maintainer}: R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

Authors:
\itemize{
  \item R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})
}

}
\keyword{internal}
